Output library to platform directory based on target info
The approach taken here is to read target_{os,arch} in build.rs,
determine the correct platform directory and filename (minus the
file extension), and then expose that information in environment
variables. These environment variables can then be read in the
xbuild wrapper script, and used to place things in the correct spot.