refactor: Refactor device logic
This is the start of the device refactor work. I've decided to keep it as minimal as possible for now.
This gives us enough of a skeleton to implement new devices that provide poses.
Devices are now stored in a Vec, accessed via the TrackedDeviceList struct.
They currently only provide a pose, and access to the device's interaction profile.

Co-authored-by: Sapphire <imsapphire0@gmail.com>